+++ src/mongo/db/query/lite_parsed_query.h
@@ -62,8 +62,8 @@ public:
      * Constructs a LiteParseQuery object as though it is from a legacy QueryMessage.
      */
     static StatusWith<std::unique_ptr<LiteParsedQuery>> makeAsOpQuery(NamespaceString nss,
-                                                                      int ntoskip,
-                                                                      int ntoreturn,
+                                                                      long long ntoskip,
+                                                                      long long ntoreturn,
                                                                       int queryoptions,
                                                                       const BSONObj& query,
                                                                       const BSONObj& proj,
@@ -78,10 +78,11 @@ public:
      * Constructs a LiteParseQuery object that can be used to serialize to find command
      * BSON object.
      */
-    static StatusWith<std::unique_ptr<LiteParsedQuery>> makeAsFindCmd(NamespaceString nss,
-                                                                      const BSONObj& query,
-                                                                      const BSONObj& sort,
-                                                                      boost::optional<int> limit);
+    static StatusWith<std::unique_ptr<LiteParsedQuery>> makeAsFindCmd(
+        NamespaceString nss,
+        const BSONObj& query,
+        const BSONObj& sort,
+        boost::optional<long long> limit);
 
     /**
      * Converts this LPQ into a find command.
@@ -159,15 +160,15 @@ public:
         return _hint;
     }
 
-    static const int kDefaultBatchSize;
+    static const long long kDefaultBatchSize;
 
     int getSkip() const {
         return _skip;
     }
-    boost::optional<int> getLimit() const {
+    boost::optional<long long> getLimit() const {
         return _limit;
     }
-    boost::optional<int> getBatchSize() const {
+    boost::optional<long long> getBatchSize() const {
         return _batchSize;
     }
     bool wantMore() const {
@@ -263,8 +264,8 @@ private:
      */
     Status validate() const;
 
-    Status init(int ntoskip,
-                int ntoreturn,
+    Status init(long long ntoskip,
+                long long ntoreturn,
                 int queryOptions,
                 const BSONObj& queryObj,
                 const BSONObj& proj,
@@ -314,8 +315,8 @@ private:
     int _skip = 0;
     bool _wantMore = true;
 
-    boost::optional<int> _limit;
-    boost::optional<int> _batchSize;
+    boost::optional<long long> _limit;
+    boost::optional<long long> _batchSize;
 
     bool _fromCommand = false;
     bool _explain = false;
